app/google_photos.py
import os
import google.auth
from googleapiclient.discovery import build
from google.oauth2.credentials import Credentials
from google.auth.transport.requests import Request
from PIL import Image

class GooglePhotos:
    def __init__(self, client_id, client_secret, redirect_uri):
        self.client_id = client_id
        self.client_secret = client_secret
        self.redirect_uri = redirect_uri
        self.credentials = self.load_credentials()
        self.service = self.build_photos_service()

    def load_credentials(self):
        creds = None
        if 'credentials' in session:
            creds = Credentials.from_authorized_user_info(session['credentials'])
        return creds

    def build_photos_service(self):
        if self.credentials and self.credentials.expired and self.credentials.refresh_token:
            self.credentials.refresh(Request())
        service = build('photoslibrary', 'v1', credentials=self.credentials)
        return service

    def download_image(self, photo_url):
        # Download the image from the provided URL (you may need to adjust this part)
        # Example: Use requests or urllib to download the image

        # For demonstration, we assume that `image_bytes` contains the image data
        # Replace this with actual code to download the image

        image_bytes = b''  # Replace with actual image data
        return image_bytes

    def process_image(self, image_id):
        # Download the image using its ID (you should implement this part)
        image_bytes = self.download_image_by_id(image_id)

        if image_bytes:
            # Process the image (flip it horizontally)
            processed_image = self.flip_image(image_bytes)

            # Upload the processed image (you should implement this part)
            upload_result = self.upload_image(processed_image)

            return upload_result

        return False

    def download_image_by_id(self, image_id):
        # Use the Google Photos API to download the image data by ID (you should implement this part)
        # Example: Use `self.service.mediaItems().get()` to retrieve the image data

        # For demonstration, we assume that `image_bytes` contains the image data
        # Replace this with actual code to download the image

        image_bytes = b''  # Replace with actual image data
        return image_bytes

    def flip_image(self, image_bytes):
        # Load the image using PIL
        image = Image.open(io.BytesIO(image_bytes))

        # Flip the image horizontally
        flipped_image = image.transpose(Image.FLIP_LEFT_RIGHT)

        # Convert the flipped image back to bytes
        flipped_image_bytes = io.BytesIO()
        flipped_image.save(flipped_image_bytes, format='JPEG')
        return flipped_image_bytes.getvalue()

    def upload_image(self, image_bytes):
        # Use the Google Photos API to upload the processed image (you should implement this part)
        # Example: Use `self.service.upload()` to upload the image

        # For demonstration, we assume that `upload_result` contains the upload result
        # Replace this with actual code to upload the image

        upload_result = {}  # Replace with actual upload result
        return upload_result

