requirements.txt
Flask
requests
google-auth
google-auth-oauthlib
google-api-python-client
Pillow
Flask-OAuthlib  # Optional, for easier OAuth handling
